Added a few workarounds to have less segfaults... still the SaveCenter does not work (at all) with plugins such as the BeatBox
git-svn-id: https://svn.code.sf.net/p/wired/code/trunk/wired@1414 6f5cd6f6-390f-0410-abd1-b47fe60d5253

+++ src/plugins/beatbox/BeatBox.cpp
@@ -1865,7 +1865,7 @@ void	WiredBeatBox::LoadChannel(SaveElement *channelData)
 
 void	WiredBeatBox::LoadBank(SaveElement *bankData)
 {
-  int			b;
+  long			b;
   SaveElementArray	patterns;
 
   bankData->getKey().AfterLast('_').ToLong((long *)&b);
@@ -1874,7 +1874,7 @@ void	WiredBeatBox::LoadBank(SaveElement *bankData)
   patterns = bankData->getChildren();
 
   for(int i = 0; i < patterns.GetCount(); i++)
-    LoadPattern(patterns[i], b);
+    LoadPattern(patterns[i], (int)b);
 }
 
 void	WiredBeatBox::LoadPattern(SaveElement *patternData, int bank)
